+++ CHANGELOG.md
@@ -10,6 +10,7 @@ This CHANGELOG follows the format listed at [Keep A Changelog](http://keepachang
 ### Changed
 - Added option -j to handler-graphite-notify.rb to allow configuration file to be defined on command line. This matches the 
   config in handler-graphite-status.rb
+- Updated sensu-plugin dependency to use a pessimistic version constraint
 
 ### Fixed
 - Correct error message for incorrect graphite expression